Move shader buffer structs out of the shaders crate
