+++ crates/shader/src/lib.rs
@@ -1,13 +1,4 @@
 mod error;
-mod glyph;
-mod path;
-mod quad;
-mod sprite;
-
-pub use glyph::*;
-pub use path::*;
-pub use quad::*;
-pub use sprite::*;
 
 use error::ErrorLogger;
