+++ src/quad.rs
@@ -1,9 +1,27 @@
+use glam::*;
 use glamour::{Point2, Size2};
 use palette::Srgba;
-use shader::{InstancedQuad, ShaderConstants, ShaderModules};
+use shader::{ShaderConstants, ShaderModules};
 use wgpu::*;
 
 use crate::{renderer::Drawable, scene::Layer, Quad, Renderer};
+
+#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable, Default)]
+#[repr(C, align(64))]
+// An axis aligned quad supporting positioning, scaling, corner radius, and optionally an internal blur with
+// the previous layer or an external blur for use with shadows.
+pub struct InstancedQuad {
+    pub color: Vec4,
+    pub _padding: Vec4,
+    pub top_left: Vec2,
+    pub size: Vec2,
+    pub __padding: Vec2,
+    pub corner_radius: f32,
+    // 0: no blur
+    // <0: internal blur of the background with kernel radius `blur`
+    // >0: external blur of quad edge with radius `blur`
+    pub blur: f32,
+}
 
 pub struct QuadState {
     buffer: Buffer,
